Add files via upload
LDBRDB is a piece of the program that finds the locations along the width of the river where the water edge (threepeak) requirement is met. It takes the difference between the two bank locations and provides the river width. The master has been updated to reflect that. The master program also has the portion of the program to find the array of pixel values along 10cm segments of the transect, to select out the midpoints of each of those segments, and to return an array that has their physical distances as the index.