The bathymetric profile of Buffalo Creek 1
Attached is the bathymetry profile. The final profile I was using is the first sheet, with the corresponding sheets having to do with the transform I used to find those final values, given the two forms of measurement (flow tracker and automatic level). I'd have to start over to ID how I got there all the way, which i will do at some point, but the first sheet is what you need.